requirements.txt
streamlit==1.24.1
langchain==0.0.348
openai==0.27.8
beautifulsoup4==4.12.2
python-dotenv==1.0.0
faiss-cpu==1.7.4
chromadb==0.3.28
requests==2.31.0
numpy==1.26.4